Create reusable CardDocument
